requirements.txt
openai==1.55.3
httpx==0.27.2
python-dotenv==1.0.0
click==8.1.7
nltk==3.8.1
jieba==0.42.1
rouge-chinese==1.0.3
pathlib==1.0.1
